smartzone 100 appliance support
im currently not switched to the vSZ so i have to handle the hardware version (SZ100). this are my current modifications to get run you script. maybe you can use this changes to improve your code. so just an merge isnt the result. so my python skills are not so hard to get an script that can handle both...